Exploit the worker thread to write into the log file

+++ msg_dumper_log.cpp
@@ -12,11 +12,17 @@
 char* MsgDumperLog::LOG_FOLDER = "log";
 
 MsgDumperLog::MsgDumperLog() :
+	pid(0),
 	exit(0),
 	log_filename(NULL),
 	log_filepath(NULL)
 {
 
+}
+
+MsgDumperLog::~MsgDumperLog()
+{
+//	deinitialize();
 }
 
 void* MsgDumperLog::msg_dumper_log_handler(void* void_ptr)
@@ -39,11 +45,32 @@ unsigned short MsgDumperLog::msg_dumper_log_handler_internal()
 	WRITE_DEBUG_SYSLOG("The worker thread of writing message is running");
 	while (!exit)
 	{
-		ret = write_logfile();
-		if (CHECK_MSG_DUMPER_FAILURE(ret))
-			break;
-//		WRITE_DEBUG_SYSLOG("Run!");
-//		sleep(1);
+		pthread_mutex_lock (&mut);
+//wait for the signal with cond as condition variable
+		pthread_cond_wait(&cond, &mut);
+
+// Move the message
+		int buffer_vector_size = buffer_vector.size();
+		if (buffer_vector_size > 0)
+		{
+			for (int i = 0 ; i < buffer_vector_size ; i++)
+			{
+				WRITE_DEBUG_FORMAT_SYSLOG(MSG_DUMPER_LONG_STRING_SIZE, "Move the message[%s] to another buffer", buffer_vector[i]);
+				char* elem =  buffer_vector[i];
+				buffer_vector[i] = NULL;
+				write_vector.push_back(elem);
+			}
+// Clean-up the container
+			buffer_vector.clear();
+		}
+		pthread_mutex_unlock (&mut);
+
+		if (write_vector.size() > 0)
+		{
+			ret = write_logfile();
+			if (CHECK_MSG_DUMPER_FAILURE(ret))
+				break;
+		}
 	}
 
 	WRITE_DEBUG_SYSLOG("The worker thread of writing message is going to die");
@@ -97,6 +124,8 @@ unsigned short MsgDumperLog::write_logfile()
 		delete[] write_vector[i];
 		write_vector[i] = NULL;
 	}
+
+// Clean-up the container
 	write_vector.clear();
 
 // Close the file
@@ -135,13 +164,19 @@ unsigned short MsgDumperLog::create_logfolder()
 unsigned short MsgDumperLog::initialize(void* config)
 {
 	int api_ret;
-// Create the workder thread
+// Create the worker thread
 	api_ret = pthread_create(&pid, NULL, msg_dumper_log_handler, this);
 	if (api_ret != 0)
 	{
 		WRITE_ERR_FORMAT_SYSLOG(MSG_DUMPER_STRING_SIZE, "pthread_create() failed, due to: %d", api_ret);
 		return MSG_DUMPER_FAILURE_UNKNOWN;
 	}
+
+	exit = 0;
+// Initialize the synchronization object
+	mut = PTHREAD_MUTEX_INITIALIZER;
+	cond = PTHREAD_COND_INITIALIZER;
+
 // Create the log folder
 	unsigned short ret = create_logfolder();
 	if (CHECK_MSG_DUMPER_FAILURE(ret))
@@ -178,24 +213,41 @@ unsigned short MsgDumperLog::write_msg(const char* msg)
 	memset(new_msg, 0x0, sizeof(char) * MSG_DUMPER_LONG_STRING_SIZE);
 	snprintf(new_msg, MSG_DUMPER_LONG_STRING_SIZE, "[%s] %s\n", time_string, msg);
 	WRITE_DEBUG_FORMAT_SYSLOG(MSG_DUMPER_LONG_STRING_SIZE, "New message: %s", new_msg);
+
+	pthread_mutex_lock(&mut);
 	buffer_vector.push_back(new_msg);
+// Wake up waiting thread with condition variable, if it is called before this function
+	pthread_cond_signal(&cond);
+	pthread_mutex_unlock(&mut);
 
 	return MSG_DUMPER_SUCCESS;
 }
 
 unsigned short MsgDumperLog::deinitialize()
 {
-	WRITE_DEBUG_SYSLOG("Try to kill the worker thread...");
-	__sync_lock_test_and_set(&exit, 1);
-
-	int api_ret;
-	char api_ret_desc[256];
-	api_ret = pthread_join(pid, (void**)&api_ret_desc);
-	if (api_ret != 0)
+	if (exit == 0)
 	{
-		WRITE_ERR_FORMAT_SYSLOG(MSG_DUMPER_STRING_SIZE, "pthread_join() failed, due to: %s", api_ret_desc);
-		return MSG_DUMPER_FAILURE_UNKNOWN;
+		WRITE_DEBUG_SYSLOG("Try to kill the worker thread...");
+		__sync_lock_test_and_set(&exit, 1);
+
+// Notify the worker thread to exit
+		pthread_mutex_lock(&mut);
+		pthread_cond_signal(&cond);
+		pthread_mutex_unlock(&mut);
+
+		int api_ret;
+		char api_ret_desc[256];
+		api_ret = pthread_join(pid, (void**)&api_ret_desc);
+		if (api_ret != 0)
+		{
+			WRITE_ERR_FORMAT_SYSLOG(MSG_DUMPER_STRING_SIZE, "pthread_join() failed, due to: %s", api_ret_desc);
+			return MSG_DUMPER_FAILURE_UNKNOWN;
+		}
+		pid = 0;
 	}
+
+    pthread_cond_destroy(&cond);
+    pthread_mutex_destroy(&mut);
 
 	if (log_filepath != NULL)
 	{
